requirements.txt
# Full ML Configuration Requirements for Multimodal Librarian
# AWS-Native implementation with Neptune and OpenSearch
# This includes all advanced ML capabilities for comprehensive demos

# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
websockets==12.0
python-multipart==0.0.6

# Security dependencies
cryptography>=41.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# PDF processing dependencies (enhanced for full ML)
PyMuPDF==1.23.8
pdfplumber==0.10.3
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3

# Full ML and embedding dependencies - compatible versions
sentence-transformers>=2.2.0
torch>=2.1.0,<2.5.0
torchvision>=0.16.0,<0.20.0
torchaudio>=2.1.0,<2.5.0
transformers>=4.35.0,<4.50.0
accelerate>=0.24.0,<1.0.0
datasets>=2.14.0,<3.0.0
evaluate>=0.4.0,<1.0.0

# Advanced NLP and ML - compatible versions
spacy>=3.7.0,<3.9.0
scikit-learn>=1.3.0,<1.6.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<2.3.0
scipy>=1.11.0,<1.15.0
nltk>=3.8.0,<4.0.0

# Database dependencies
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# Environment and configuration
python-dotenv>=1.0.1,<2.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0
rich==13.7.0
psutil==5.9.6

# External API dependencies - compatible versions
google-generativeai>=0.3.0,<1.0.0
requests==2.31.0

# Full multimedia generation dependencies - compatible versions
matplotlib>=3.8.0,<3.9.0
seaborn>=0.12.0,<0.14.0
plotly>=5.17.0,<6.0.0
gTTS>=2.4.0,<3.0.0
pyttsx3>=2.90,<3.0.0
moviepy>=1.0.3,<2.0.0
opencv-python>=4.8.0,<5.0.0

# Export functionality dependencies
python-docx==1.1.0
reportlab==4.0.7
python-pptx==0.6.23
openpyxl==3.1.2

# AWS-Native database dependencies
gremlinpython>=3.6.0,<4.0.0  # For Neptune Gremlin queries
opensearch-py>=2.4.0,<3.0.0  # For OpenSearch operations
requests-aws4auth>=1.2.0,<2.0.0  # For AWS IAM authentication
networkx==3.2.1  # For graph operations

# Local development database dependencies
neo4j>=5.12.0,<6.0.0  # For local Neo4j graph database
pymilvus>=2.3.0,<3.0.0  # For local Milvus vector database



# AWS dependencies
boto3==1.34.0
botocore==1.34.0

# Testing dependencies (for ML validation)
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
httpx==0.25.2

# Development dependencies
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Additional ML utilities
joblib>=1.3.0,<2.0.0
tqdm>=4.66.0,<5.0.0
python-dateutil==2.8.2
pytz==2023.3
aiofiles>=23.2.0,<24.0.0
docker>=6.1.0,<7.0.0

# Image processing and computer vision
imageio>=2.31.0,<3.0.0
# Additional dependencies found by dependency scanner
aiohttp>=3.9.0,<4.0.0
celery>=5.3.0,<6.0.0
PyYAML>=6.0.0,<7.0.0
moto>=4.2.0,<5.0.0
starlette>=0.27.0,<1.0.0
urllib3>=1.26.0,<3.0.0
jinja2>=3.1.0
